fix(settings): change type address menu location
